use `https.get` instead of native `fetch`
